requirements.txt
streamlit
faiss-cpu
sentence-transformers
groq
pypdf
numpy
python-dotenv